+++ Game/Game/CPlayer.h
@@ -23,7 +23,7 @@ class CPlayer : public CPhysicsBody, public CGameObject
 public:
 	static std::set<CPlayer*> playersInWorld;
 
-	int maxHealth;
+	float maxHealth;
 	float moveSpeed;
 	float coolDown;
 	bool isPlayerOne;
@@ -45,7 +45,7 @@ public:
 	// health info
 	sf::Text healthText;
 	sf::String healthString;
-	int health;
+	float health;
 
 	CPlayer(sf::Keyboard::Key _up, sf::Keyboard::Key _down, sf::Keyboard::Key _left, sf::Keyboard::Key _right, sf::Keyboard::Key _shoot, sf::Vector2f _spawnPos, bool _isPlayerOne);
 	~CPlayer();
